pyproject.toml
[project]
name = "businessanalyzer"
version = "0.1.0"
description = "Investment Decision Support System with AI-powered sentiment analysis"
requires-python = ">=3.14"
dependencies = [
    # Web Framework
    "fastapi>=0.123.0",
    "uvicorn[standard]>=0.38.0",
    "python-multipart>=0.0.6",
    # Database (NO ORM - plain SQL only)
    "mysql-connector-python>=8.0.33",
    # Security
    "bcrypt>=4.0.1",
    "python-jose[cryptography]>=3.3.0",
    # Task Queue (RabbitMQ)
    "celery>=5.3.0",
    "kombu>=5.3.0",
    # Web Scraping
    "beautifulsoup4>=4.12.0",
    "scrapy>=2.11.0",
    "requests>=2.31.0",
    "lxml>=5.0.0",
    # AI/ML
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    # Template Rendering
    "jinja2>=3.1.2",
    # Environment Variables
    "python-dotenv>=1.0.0",
    # Data Processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    # HTTP Client
    "httpx>=0.24.0",
    # Validation
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "itsdangerous>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
